requirements.lock.txt
absl-py==2.3.1
accelerate==1.10.1
accelerated-scan==0.2.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.16.5
aniso8601==10.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asciitree==0.3.3
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
av==15.1.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.1
bitsandbytes==0.46.0
black==24.10.0
blinker==1.9.0
boto3==1.40.37
botocore==1.40.37
braceexpand==0.1.7
bracex==2.6
catalogue==2.0.10
cdifflib==1.2.6
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
clip==0.2.0
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.9.0
contextlib2==21.6.0
contourpy==1.3.3
coverage==7.10.7
cryptography==42.0.8
cuda-bindings==13.0.1
cuda-pathfinder==1.2.3
cuda-python==13.0.1
cycler==0.12.1
cytoolz==1.0.1
DataProperty==1.1.0
datasets==4.1.1
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.35.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.21.2
editdistance==0.8.1
einops==0.8.1
einops-exts==0.0.4
evaluate==0.4.6
exceptiongroup==1.3.0
executing==2.2.1
fabric==3.2.2
faiss-cpu==1.12.0
fastapi==0.117.1
fasteners==0.20
ffmpeg-python==0.2.0
fiddle==0.3.0
filelock==3.19.1
Flask==3.1.2
Flask-RESTful==0.3.10
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.35.1
hydra-core==1.3.2
hyperframe==6.1.0
HyperPyYAML==1.2.2
idna==3.10
ijson==3.4.0
imageio==2.37.0
imagesize==1.4.1
immutabledict==4.2.0
importlib_metadata==8.7.0
indic_numtowords==1.1.0
inflect==7.5.0
iniconfig==2.1.0
inquirerpy==0.3.4
intervaltree==3.1.0
invoke==2.2.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isort==5.13.2
itsdangerous==2.2.0
Janome==0.5.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.0
jiwer==3.1.0
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
kaldi-python-io==1.2.2
kaldialign==0.9.1
kiwisolver==1.4.9
kornia==0.8.1
kornia_rs==0.1.9
langdetect==1.0.9
lark==1.3.0
latexcodec==3.0.1
lazy_loader==0.4
ledoc-ui==0.1.0
leptonai==0.26.6
Levenshtein==0.27.1
lhotse==1.32.2
libcst==1.8.4
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.2
lilcom==1.8.1
llvmlite==0.45.0
loguru==0.7.3
lxml==6.0.2
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
mediapy==1.1.6
megatron-core==0.13.1
megatron-energon==5.2.0
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multi-storage-client==0.29.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@3bfbcb676de9224c1a3404dbc2a27f7d756c8dd5
nemo_run==0.5.0
nemo_text_processing==1.1.0
nerfacc==0.5.3
networkx==3.5
ninja==1.13.0
nltk==3.9.1
num2words==0.5.14
numba==0.62.0
numba-cuda==0.15.1
numcodecs==0.15.1
numexpr==2.12.1
numpy==1.26.4
nv-one-logger-core==2.3.1
nv-one-logger-pytorch-lightning-integration==2.3.1
nv-one-logger-training-telemetry==2.3.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-modelopt==0.35.0
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
nvidia-resiliency-ext==0.4.1
nvidia_lm_eval==25.6.1
nvtx==0.2.13
omegaconf==2.3.0
onnx==1.19.0
open-clip-torch==2.24.0
openai==1.109.0
OpenCC==1.1.9
opentelemetry-api==1.37.0
optuna==4.5.0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pangu==4.0.6.1
parameterized==0.9.0
paramiko==4.0.0
parso==0.8.5
pathspec==0.12.1
pathvalidate==3.3.1
peft==0.17.1
pesq==0.0.4
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
plac==1.4.5
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
portalocker==3.2.0
prettytable==3.16.0
primePy==1.3
progress==1.6.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==5.29.5
psutil==7.1.0
ptyprocess==0.7.0
PuLP==3.3.0
pure_eval==0.2.3
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==21.0.0
pybind11==3.0.1
pybtex==0.25.1
pybtex-docutils==1.0.3
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
PyNaCl==1.6.0
pynini==2.1.6.post1
pynvml==13.0.1
pyopenjtalk==0.4.1
pyparsing==3.2.5
pypinyin==0.55.0
pypinyin-dict==0.9.0
pyre-extensions==0.0.32
PySocks==1.7.1
pystoi==0.4.1
pytablewriter==1.2.1
pytest==8.4.2
pytest-mock==3.15.1
pytest-runner==6.0.1
pytest_httpserver==1.1.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.14
RapidFuzz==3.14.1
ray==2.49.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
resampy==0.4.3
rich==14.1.0
roman-numerals-py==3.1.0
rouge_score==0.1.2
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3fs==0.4.2
s3transfer==0.14.0
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semver==3.0.4
sentence-transformers==5.1.1
sentencepiece==0.2.1
sentry-sdk==2.38.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
sox==1.5.0
soxr==1.0.0
speechbrain==1.0.3
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.5
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
StrEnum==0.4.15
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
taming-transformers==0.0.1
tcolorpy==0.1.7
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorstore==0.1.71
termcolor==3.1.0
text-unidecode==1.3
textdistance==4.6.3
texterrors==1.0.9
threadpoolctl==3.6.0
tiktoken==0.7.0
timm==1.0.20
tokenizers==0.22.2
toml==0.10.2
toolz==1.0.0
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchdiffeq==0.2.5
torchmetrics==1.8.2
torchprofile==0.0.4
torchsde==0.2.6
torchvision==0.23.0
torchx==0.7.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
trampoline==0.1.2
transformers==4.57.6
tree-sitter==0.25.1
tree-sitter-python==0.25.0
trimesh==4.8.2
triton==3.4.0
typeguard==4.4.4
typepy==1.3.4
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.37.0
wandb==0.22.0
wcmatch==10.1
wcwidth==0.2.14
webdataset==1.0.2
Werkzeug==3.1.3
wget==3.2
whisper_normalizer==0.1.12
word2number==1.1
wrapt==1.17.3
xattr==1.2.0
xxhash==3.5.0
yarl==1.20.1
zarr==2.18.7
zipp==3.23.0
zstandard==0.25.0
chromadb
dotenv
sqlalchemy>=2.0
aiosqlite>=0.19